Bouge le code du cron pour synchro les users et les gestionnaires dans server
