+++ src/server/cron/launch.ts
@@ -1,7 +1,7 @@
 import * as Sentry from '@sentry/node';
 
+import { upsertUsersFromGestionnaireSheet } from '@/server/services/airtable';
 import { dailyNewManagerMail, dailyRelanceMail, weeklyOldManagerMail } from '@/server/services/manager';
-import { updateUsers } from '@/services/users';
 
 import '@root/sentry.node.config';
 
@@ -9,7 +9,7 @@ export const jobs: Record<string, any> = {
   dailyNewManagerMail,
   weeklyOldManagerMail,
   dailyRelanceMail,
-  updateUsers,
+  upsertUsersFromGestionnaireSheet,
 };
 
 export const launchJob = async (job: string) => {
